chore(core, common): move `Iterate` and `IsEmpty` to `core`
