+++ src/plugins/animations/src/systems/set_directional_animation_weights.rs
@@ -112,12 +112,12 @@ mod tests {
 	use crate::components::animation_dispatch::AnimationPlayerOf;
 	use common::traits::{
 		handles_animations::{Animation, AnimationKey, Directional},
-		iterate::Iterate,
 		wrap_handle::{GetHandle, WrapHandle},
 	};
 	use std::{collections::HashMap, slice::Iter};
 	use test_case::test_case;
 	use testing::{SingleThreadedApp, assert_eq_approx, new_handle};
+	use zyheeda_core::prelude::*;
 
 	#[derive(Component)]
 	struct _Dispatch {
