+++ src/plugins/animations/src/systems/set_pitch_animation_weights.rs
@@ -120,13 +120,13 @@ mod tests {
 				PitchedForward,
 				SkillAnimation,
 			},
-			iterate::Iterate,
 			wrap_handle::{GetHandle, WrapHandle},
 		},
 	};
 	use std::{collections::HashMap, slice::Iter};
 	use test_case::test_case;
 	use testing::{SingleThreadedApp, assert_eq_approx, new_handle};
+	use zyheeda_core::prelude::*;
 
 	#[derive(Component)]
 	struct _Dispatch {
