+++ src/plugins/common/src/traits.rs
@@ -31,9 +31,7 @@ pub mod handles_saving;
 pub mod handles_skill_physics;
 pub mod init_resource;
 pub mod insert;
-pub mod is_empty;
 pub mod iter_descendants_conditional;
-pub mod iterate;
 pub mod iteration;
 pub mod load_asset;
 pub mod load_folder_assets;
