+++ src/plugins/input/src/resources/key_map.rs
@@ -8,7 +8,6 @@ use common::{
 	traits::{
 		handles_custom_assets::TryLoadFrom,
 		handles_input::{GetInput, InvalidUserInput, UpdateKey},
-		iterate::Iterate,
 		iteration::IterFinite,
 		load_asset::LoadAsset,
 		thread_safe::ThreadSafe,
