+++ src/plugins/loadout/src/system_parameters/loadout/read/skills.rs
@@ -15,9 +15,9 @@ use common::{
 			skills::{ReadSkills, SkillIcon, SkillToken, Skills},
 		},
 		handles_localization::Token,
-		iterate::Iterate,
 	},
 };
+use zyheeda_core::prelude::*;
 
 impl GetContext<Skills> for LoadoutReader<'static, 'static> {
 	type TContext<'ctx> = SkillsView<'ctx>;
