+++ src/plugins/loadout/src/system_parameters/loadout_activity/read/active_skills.rs
@@ -2,10 +2,8 @@ use crate::{
 	components::queue::Queue,
 	system_parameters::loadout_activity::LoadoutActivityReadContext,
 };
-use common::traits::{
-	handles_loadout::{ActiveSkill, ActiveSkills},
-	iterate::Iterate,
-};
+use common::traits::handles_loadout::{ActiveSkill, ActiveSkills};
+use zyheeda_core::prelude::*;
 
 impl ActiveSkills for LoadoutActivityReadContext<'_> {
 	type TIter<'a>
