+++ src/zyheeda_core/src/collections/is_empty.rs
