+++ src/zyheeda_core/src/collections/iterate/slice.rs
