+++ src/zyheeda_core/src/collections/iterate/vec.rs
